third_party/WebKit/LayoutTests/http/tests/security/xssAuditor/script-tag-with-source-expected.txt

CONSOLE ERROR: line 4: The XSS Auditor refused to execute a script in 'http://localhost:8000/security/xssAuditor/resources/echo-intertag.pl?q=%3Cscript%20src=%27http://127.0.0.1:8000/security/xssAuditor/resources/xss.js%27%3E%3C/script%3E' because its source code was found within the request. The server sent an 'X-XSS-Protection' header requesting this behavior.
